all for today, we have moved to my own canvas still we have one major bug, i'll try to find out what's going on later

+++ components/graphics/canvas.h
@@ -34,7 +34,7 @@ public:
 
     RGBColor *get_pixel_ptr();
 
-    size_t pitch() const;
+    int pitch() const;
 
 protected:
